Fix a warning related to printing line numbers

+++ src/CppUTest/TestRegistry.cpp
@@ -135,7 +135,7 @@ void TestRegistry::listTestLocations(TestResult& result)
             testLocation += ".";
             testLocation += test->getFile();
             testLocation += ".";
-            testLocation += StringFromFormat("%d\n",test->getLineNumber());
+            testLocation += StringFromFormat("%d\n",(int) test->getLineNumber());
 
             testLocations += testLocation;
     }
